refactor: migrated event permission checks from EventController to RestrictToEventOwner middleware, applying it to edit, update, and destroy routes in web.php. Removed redundant children_affected validation from controller, retaining it in the Event model

+++ routes/web.php
@@ -52,9 +52,9 @@ Route::middleware(['auth', 'role:admin|user'])->prefix('admin')->group(function 
     Route::get('/events', [EventController::class, 'index'])->name('admin.events.index');
     Route::get('/events/create', [EventController::class, 'create'])->name('admin.events.create');
     Route::post('/events', [EventController::class, 'store'])->name('admin.events.store');
-    Route::get('/events/{event}/edit', [EventController::class, 'edit'])->name('admin.events.edit');
-    Route::put('/events/{event}', [EventController::class, 'update'])->name('admin.events.update');
-    Route::delete('/events/{event}', [EventController::class, 'destroy'])->name('admin.events.destroy');
+    Route::get('/events/{event}/edit', [EventController::class, 'edit'])->name('admin.events.edit')->middleware('event.owner');
+    Route::put('/events/{event}', [EventController::class, 'update'])->name('admin.events.update')->middleware('event.owner');
+    Route::delete('/events/{event}', [EventController::class, 'destroy'])->name('admin.events.destroy')->middleware('event.owner');
 });
 
 Route::middleware(['auth', 'role:admin|user'])->group(function () {
